+++ source/notify_sonarr/changelog.md
@@ -1,3 +1,6 @@
 
+**<span style="color:#56adda">0.0.2</span>**
+- Add ability to import episode into Sonarr on task complete
+
 **<span style="color:#56adda">0.0.1</span>**
 - Initial version